Fix bugs
- Previously renamed function in `CombatEntity` was being called by the old name, resulting in a compiler error
- Move `PlayerHPDecrease` component into the `PlayerCharacter` namespace and drop the unnecessary noun suffix
- Add `PlayerHPDecrease` component to the player character game object in the scene
- Create additional sorting layers to fix sprite sorting issues
- Fix that the unlock sound would play even after the generator is already unlocked
- Add value checks to fix negative damages when the defense is larger than the incoming attack
