+++ Assets/Scripts/Generators/Generator.cs
@@ -17,24 +17,19 @@ namespace Generators {
         public UnityEvent onActivated;
 
         private GameObject keyUI;
-        private bool isActivated;
+
+        private bool isUnlocked; // unlocked with the key
+        private bool isActivated; // activated by the puzzle or key
 
         public void OnPlayerCollide(GameObject player)
         {
-            if (isActivated)
-                return; // do nothing if this is already activated
+            if (isActivated || isUnlocked)
+                return; // do nothing if this is already activated or unlocked
 
             if (!inventory.CanActivate(this))
                 return; // do nothing if the player does not have the key
 
-            if (keyUI != null) // clear out the key in the ui
-            {
-                Destroy(keyUI);
-                keyUI = null;
-            }
-
-            if (unlockSoundPlayer != null)
-                unlockSoundPlayer.Play();
+            Unlock();
 
             if (puzzle == null)
                 Activate(); // activate if it does not have a puzzle associated with it
@@ -54,6 +49,20 @@ namespace Generators {
         public void SetKeyUI(GameObject ui)
         {
             keyUI = ui;
+        }
+
+        private void Unlock()
+        {
+            isUnlocked = true;
+
+            if (keyUI != null) // clear out the key in the ui
+            {
+                Destroy(keyUI);
+                keyUI = null;
+            }
+
+            if (unlockSoundPlayer != null)
+                unlockSoundPlayer.Play();
         }
     }
 }
