+++ ProjectSettings/TagManager.asset
@@ -44,3 +44,9 @@ TagManager:
   - name: Default
     uniqueID: 0
     locked: 0
+  - name: Player
+    uniqueID: 2434974901
+    locked: 0
+  - name: Emotes
+    uniqueID: 1226726929
+    locked: 0
